requirements.txt
apache-airflow-providers-oracle==3.11.0
apache-airflow-providers-postgres==5.13.1
astronomer-cosmos==1.8.0
astro-sdk-python==1.8.0
openpyxl==3.1.2
dbt-core==1.8.4
dbt-oracle==1.8.3
dbt-postgres==1.8.2
unidecode==1.3.8
pyxlsb==1.0.10
psycopg2-binary==2.9.9